requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
pydantic==2.11.4
passlib[bcrypt]==1.7.4
loguru==0.7.3
confluent-kafka==2.10.0
python-dotenv==1.1.0
pydantic-settings==2.9.1
streamlit==1.45.0
streamlit-autorefresh==1.0.1
plotly==6.0.1
